--- src/containers/AccessPointDetails/components/index.module.scss
@@ -1,76 +1,0 @@
-:global(.ant-card) {
-  margin-top: 10px;
-  border: 1px solid #434343;
-  border-radius: 5px;
-}
-
-:global(.ant-tag) {
-  border: 1px solid #434343;
-  border-radius: 5px;
-}
-
-:global(.ant-alert) {
-  border: 1px solid #434343;
-  border-radius: 5px;
-  width: 100%;
-  margin-right: 20px;
-}
-
-:global(.ant-textarea) {
-  border: 1px solid #434343;
-  border-radius: 5px;
-  margin-top: 10px;
-}
-
-:global(.ant-collapse) {
-  margin-top: 10px;
-  background: #141414;
-}
-
-:global(.ant-progress) {
-  margin-right: 5px;
-}
-
-.Field {
-  background: #141414;
-  border: 1px solid #434343;
-  border-radius: 5px;
-  margin-bottom: 5px;
-}
-:global(.ant-form-item) {
-  flex: 1;
-}
-
-.InlineDiv {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-}
-
-.InlineDiv > * {
-  margin: 0 5px;
-}
-
-.InlineBetweenDiv {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.saveButton {
-  border: 1px solid #434343;
-  border-radius: 5px;
-  margin-right: 25px;
-  margin-top: 10px;
-  width: 139px;
-}
-
-.InlineEndDiv {
-  display: flex;
-  justify-content: flex-end;
-}
-
-p {
-  font-weight: bold !important;
-  font-size: 16px;
-}
